MainFragment

+++ app/src/main/java/com/senla/fitnessapp/data/network/models/saveTrackRequest/SaveTrackRequest.kt
@@ -1,6 +1,6 @@
 package com.senla.fitnessapp.data.network.models.saveTrackRequest
 
 data class SaveTrackRequest(
-    val token: String, val id: Int? = null, val beginsAt: Long, val time: Int,
+    val token: String, val id: Int? = null, val beginsAt: Long, val time: Long,
     val distance: Int, val points: List<Point>
 )
